refactor(single-stream-instance-segmenter): keep the example in one file
The encoder lived in its own header so the C++ unit test could include it,
which split the example across two files and pushed a link dependency onto
every example's unit-test target. An example is read as one file, and the
Python implementation already keeps everything in main.py.

The C++ encoder loses its unit tests, since a test cannot include main.cpp.
Python keeps equivalent coverage, and the shared cmake change reverts with
the header.
